refactor: move pages/_app.tsx and pages/_document.tsx to app/layout.tsx and app/page.tsx
