+++ app/layout.tsx
@@ -1,24 +1,32 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "./globals.css";
+"use client";
+
+import { CssBaseline, createTheme } from "@mui/material";
 import StyledComponentsRegistry from "lib/registry";
+import { Inter } from "next/font/google";
+import Head from "next/head";
+import { ThemeProvider } from "styled-components";
+import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const theme = createTheme();
+
   return (
     <html lang="en">
+      <Head>
+        <meta name="viewport" content="initial-scale=1, width=device-width" />
+      </Head>
+
       <StyledComponentsRegistry>
-        <body className={inter.className}>{children}</body>
+        <ThemeProvider theme={theme}>
+          <CssBaseline />
+          <body className={inter.className}>{children}</body>
+        </ThemeProvider>
       </StyledComponentsRegistry>
     </html>
   );
