+++ app/page.tsx
@@ -1,0 +1,11 @@
+import { Metadata } from "next";
+import Home from "./home";
+
+export const metadata: Metadata = {
+  title: "MoonStcok",
+  description: "Good tools for investor",
+};
+
+export default async function Page() {
+  return <Home />;
+}
